requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
aiosqlite==0.19.0
APScheduler==3.10.4
PyYAML==6.0.1
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.2.1
cachetools==5.3.3
cloudscraper==1.2.71
coloredlogs==15.0.1
aiocache==0.12.2
